@@ -1,8 +1,8 @@
 [package]
 name = "bevy_mod_scripting"
-version = "0.15.0"
+version = "0.15.1"
 authors = ["Maksymilian Mozolewski <makspl17@gmail.com>"]
-edition = "2021"
+edition = "2024"
 license = "MIT OR Apache-2.0"
 description = "Multi language scripting in Bevy"
 repository = "https://github.com/makspll/bevy_mod_scripting"
