@@ -1,7 +1,7 @@
 [package]
 name = "bevy_mod_scripting_derive"
-version = "0.15.0"
-edition = "2021"
+version = "0.15.1"
+edition = "2024"
 authors = ["Maksymilian Mozolewski <makspl17@gmail.com>"]
 license = "MIT OR Apache-2.0"
 description = "Necessary functionality for Lua support with bevy_mod_scripting"
@@ -12,9 +12,9 @@ categories = ["game-development"]
 readme = "readme.md"
 
 [dependencies]
-syn = { version = "2", features = ["full"] }
-proc-macro2 = "1"
-quote = "1"
+syn = { workspace = true, features = ["full"] }
+proc-macro2 = { workspace = true }
+quote = { workspace = true }
 
 [lib]
 proc-macro = true
