@@ -1,7 +1,7 @@
 [package]
 name = "mdbook_lad_preprocessor"
 version = "0.2.0"
-edition = "2021"
+edition = "2024"
 authors = ["Maksymilian Mozolewski <makspl17@gmail.com>"]
 license = "MIT OR Apache-2.0"
 description = "Language Agnostic Declaration (LAD) file format for the bevy_mod_scripting crate"
@@ -13,17 +13,17 @@ include = ["readme.md", "/src"]
 readme = "readme.md"
 
 [dependencies]
-clap = "4"
-mdbook = "0.4"
-ladfile = { path = "../../ladfile", version = "0.5.0" }
-env_logger = "0.11"
-log = "0.4"
-serde_json = "1.0"
-regex = "1.11"
+env_logger = { workspace = true }
+clap = { workspace = true }
+ladfile = { workspace = true }
+mdbook = { workspace = true, features = ["search"] }
+log = { workspace = true }
+serde_json = { workspace = true }
+regex = { workspace = true }
 
 [dev-dependencies]
-assert_cmd = "2.0"
-pretty_assertions = "1.4.1"
+assert_cmd = { workspace = true }
+pretty_assertions = { workspace = true }
 
 [lints]
 workspace = true
