@@ -1,8 +1,8 @@
 [package]
 name = "bevy_mod_scripting_rhai"
-version = "0.15.0"
+version = "0.15.1"
 authors = ["Maksymilian Mozolewski <makspl17@gmail.com>"]
-edition = "2021"
+edition = "2024"
 license = "MIT OR Apache-2.0"
 description = "Necessary functionality for Rhai support with bevy_mod_scripting"
 repository = "https://github.com/makspll/bevy_mod_scripting"
@@ -16,11 +16,14 @@ name = "bevy_mod_scripting_rhai"
 path = "src/lib.rs"
 
 [dependencies]
-bevy = { workspace = true, default-features = false }
-rhai = { version = "1.21" }
+bevy_ecs = { workspace = true, default-features = false, features = [] }
+bevy_asset = { workspace = true, default-features = false, features = [] }
+bevy_app = { workspace = true, default-features = false, features = [] }
+bevy_log = { workspace = true, default-features = false, features = [] }
+rhai = { workspace = true, features = ["std"] }
 bevy_mod_scripting_core = { workspace = true, features = ["rhai_impls"] }
-strum = { version = "0.26", features = ["derive"] }
-parking_lot = "0.12.1"
+strum = { workspace = true, features = ["derive"] }
+parking_lot = { workspace = true }
 
 [lints]
 workspace = true
