@@ -1,21 +1,20 @@
 //! An allocator used to control the lifetime of allocations
 
-use bevy::prelude::Resource;
-use bevy::{
-    app::{App, Plugin, PostUpdate},
-    diagnostic::{Diagnostic, DiagnosticPath, Diagnostics, RegisterDiagnostic},
-	ecs::system::Res,
-	platform::collections::HashMap,
-    prelude::ResMut,
-    reflect::PartialReflect,
+use ::{
+    bevy_app::{App, Plugin, PostUpdate},
+    bevy_diagnostic::{Diagnostic, DiagnosticPath, Diagnostics, RegisterDiagnostic},
+    bevy_ecs::system::Res,
+    bevy_reflect::PartialReflect,
 };
+use bevy_ecs::{resource::Resource, system::ResMut};
+use bevy_platform::collections::HashMap;
 use parking_lot::{RwLock, RwLockReadGuard, RwLockWriteGuard};
 use std::{
     cell::UnsafeCell,
     cmp::Ordering,
     fmt::{Display, Formatter},
     hash::Hasher,
-    sync::{atomic::AtomicU64, Arc},
+    sync::{Arc, atomic::AtomicU64},
 };
 
 /// The path used for the total number of allocations diagnostic
@@ -121,7 +120,7 @@ impl ReflectAllocation {
     /// # Safety
     /// - Must only be done if no other references to this allocation exist at the same time
     pub unsafe fn take(self) -> Box<dyn PartialReflect> {
-        std::mem::transmute(self.0)
+        unsafe { std::mem::transmute(self.0) }
     }
 }
 
