@@ -6,18 +6,20 @@ use crate::asset::Language;
 use crate::bindings::function::arg_meta::ArgMeta;
 use crate::docgen::info::{FunctionInfo, GetFunctionInfo};
 use crate::{
+    ScriptValue,
     bindings::{ThreadWorldContainer, WorldContainer, WorldGuard},
     error::InteropError,
-    ScriptValue,
 };
-use bevy::platform::collections::HashMap;
-use bevy::prelude::{Reflect, Resource};
+use bevy_ecs::prelude::Resource;
+use bevy_platform::collections::HashMap;
+use bevy_reflect::Reflect;
 use parking_lot::{RwLock, RwLockReadGuard, RwLockWriteGuard};
 use std::borrow::Cow;
 use std::collections::VecDeque;
 use std::hash::Hash;
 use std::ops::{Deref, DerefMut};
 use std::sync::Arc;
+
 #[diagnostic::on_unimplemented(
     message = "This function does not fulfil the requirements to be a script callable function. All arguments must implement the ScriptArgument trait and all return values must implement the ScriptReturn trait",
     note = "If you're trying to return a non-primitive type, you might need to use Val<T> Ref<T> or Mut<T> wrappers"
@@ -626,10 +628,11 @@ variadics_please::all_tuples!(impl_script_function, 0, 13, T);
 
 #[cfg(test)]
 mod test {
-	use super::*;
+    use super::*;
+    use bevy_ecs::{prelude::Component, world::World};
 
-	fn with_local_world<F: Fn()>(f: F) {
-        let mut world = bevy::prelude::World::default();
+    fn with_local_world<F: Fn()>(f: F) {
+        let mut world = World::default();
         WorldGuard::with_static_guard(&mut world, |world| {
             ThreadWorldContainer.set_world(world).unwrap();
             f()
@@ -693,7 +696,7 @@ mod test {
 
     #[test]
     fn test_interrupted_call_releases_access_scope() {
-        #[derive(bevy::prelude::Component, Reflect)]
+        #[derive(Component, Reflect)]
         struct Comp;
 
         let fn_ = |_a: crate::bindings::function::from::Mut<Comp>| 0usize;
