@@ -1,17 +1,19 @@
 //! This module contains the [`GetTypeDependencies`] trait and its implementations for various types.
 
 use super::{
-	from::{Mut, Ref, Union, Val},
-	script_function::FunctionCallContext, DynamicScriptFunction, DynamicScriptFunctionMut,
+    DynamicScriptFunction, DynamicScriptFunctionMut,
+    from::{Mut, Ref, Union, Val},
+    script_function::FunctionCallContext,
 };
 use crate::{
-	bindings::{ReflectReference, ScriptValue}, error::InteropError}
-;
-use bevy::reflect::{FromReflect, GetTypeRegistration, TypeRegistry, Typed};
+    bindings::{ReflectReference, ScriptValue},
+    error::InteropError,
+};
 use bevy_mod_scripting_derive::impl_get_type_dependencies;
-use std::{collections::HashMap, ffi::OsString, hash::Hash, path::PathBuf};
-
-
+use bevy_platform::collections::HashMap;
+use bevy_reflect::{FromReflect, GetTypeRegistration, TypeRegistry, Typed};
+use std::collections::HashMap as StdHashMap;
+use std::{ffi::OsString, hash::Hash, path::PathBuf};
 
 macro_rules! impl_get_type_dependencies_primitives {
     ($($ty:ty),*) => {
@@ -26,11 +28,30 @@ macro_rules! impl_get_type_dependencies_primitives {
 }
 
 impl_get_type_dependencies_primitives!(
-    i8, i16, i32, i64, i128, u8, u16, u32, u64, u128, usize, isize, f32, f64, bool,
-    ScriptValue, DynamicScriptFunction, DynamicScriptFunctionMut, InteropError,
-    String, PathBuf, OsString, char
+    i8,
+    i16,
+    i32,
+    i64,
+    i128,
+    u8,
+    u16,
+    u32,
+    u64,
+    u128,
+    usize,
+    isize,
+    f32,
+    f64,
+    bool,
+    ScriptValue,
+    DynamicScriptFunction,
+    DynamicScriptFunctionMut,
+    InteropError,
+    String,
+    PathBuf,
+    OsString,
+    char
 );
-
 
 impl GetTypeDependencies for () {
     type Underlying = ();
@@ -58,75 +79,91 @@ pub trait GetTypeDependencies {
 
 impl_get_type_dependencies!(
     #[derive(GetTypeDependencies)]
-    #[get_type_dependencies(bms_core_path="crate")]
-    struct HashMap<K,V> where 
+    #[get_type_dependencies(bms_core_path = "crate")]
+    struct HashMap<K, V>
+    where
         K::Underlying: FromReflect + Eq + Hash + Typed,
-        V::Underlying: FromReflect + Typed {}
+        V::Underlying: FromReflect + Typed, {}
 );
 
 impl_get_type_dependencies!(
     #[derive(GetTypeDependencies)]
-    #[get_type_dependencies(bms_core_path="crate")]
-    struct Result<T, E> where 
+    #[get_type_dependencies(bms_core_path = "crate")]
+    struct StdHashMap<K, V>
+    where
+        K::Underlying: FromReflect + Eq + Hash + Typed,
+        V::Underlying: FromReflect + Typed, {}
+);
+
+impl_get_type_dependencies!(
+    #[derive(GetTypeDependencies)]
+    #[get_type_dependencies(bms_core_path = "crate")]
+    struct Result<T, E>
+    where
         T::Underlying: FromReflect + Typed,
-        E::Underlying: FromReflect + Typed {}
+        E::Underlying: FromReflect + Typed, {}
 );
 
 impl_get_type_dependencies!(
     #[derive(GetTypeDependencies)]
-    #[get_type_dependencies(bms_core_path="crate")]
-    struct Option<T> where 
-        T::Underlying: FromReflect + Typed {}
+    #[get_type_dependencies(bms_core_path = "crate")]
+    struct Option<T>
+    where
+        T::Underlying: FromReflect + Typed, {}
 );
 
 impl_get_type_dependencies!(
     #[derive(GetTypeDependencies)]
-    #[get_type_dependencies(bms_core_path="crate")]
-    struct Vec<T> where 
-        T::Underlying: FromReflect + Typed {}
+    #[get_type_dependencies(bms_core_path = "crate")]
+    struct Vec<T>
+    where
+        T::Underlying: FromReflect + Typed, {}
 );
-
 
 impl_get_type_dependencies!(
     #[derive(GetTypeDependencies)]
-    #[get_type_dependencies(bms_core_path="crate", underlying="Result<T1::Underlying,T2::Underlying>")]
-    struct Union<T1, T2> where 
+    #[get_type_dependencies(
+        bms_core_path = "crate",
+        underlying = "Result<T1::Underlying,T2::Underlying>"
+    )]
+    struct Union<T1, T2>
+    where
         T1::Underlying: FromReflect + Typed,
-        T2::Underlying: FromReflect + Typed {}
+        T2::Underlying: FromReflect + Typed, {}
 );
 
 impl_get_type_dependencies!(
     #[derive(GetTypeDependencies)]
-    #[get_type_dependencies(bms_core_path="crate", underlying="T", dont_recurse)]
+    #[get_type_dependencies(bms_core_path = "crate", underlying = "T", dont_recurse)]
     struct Val<T> {}
 );
 
 impl_get_type_dependencies!(
     #[derive(GetTypeDependencies)]
-    #[get_type_dependencies(bms_core_path="crate", underlying="T", dont_recurse)]
+    #[get_type_dependencies(bms_core_path = "crate", underlying = "T", dont_recurse)]
     struct Ref<'a, T> {}
 );
 
 impl_get_type_dependencies!(
     #[derive(GetTypeDependencies)]
-    #[get_type_dependencies(bms_core_path="crate", underlying="T", dont_recurse)]
+    #[get_type_dependencies(bms_core_path = "crate", underlying = "T", dont_recurse)]
     struct Mut<'a, T> {}
 );
 
 impl_get_type_dependencies!(
     #[derive(GetTypeDependencies)]
-    #[get_type_dependencies(bms_core_path="crate")]
-    struct ReflectReference where {}
+    #[get_type_dependencies(bms_core_path = "crate")]
+    struct ReflectReference {}
 );
 
 impl_get_type_dependencies!(
     #[derive(GetTypeDependencies)]
-    #[get_type_dependencies(bms_core_path="crate")]
-    struct FunctionCallContext where {}
+    #[get_type_dependencies(bms_core_path = "crate")]
+    struct FunctionCallContext {}
 );
 
-
-impl <T, const N: usize> GetTypeDependencies for [T; N] where 
+impl<T, const N: usize> GetTypeDependencies for [T; N]
+where
     T: GetTypeDependencies,
     T::Underlying: FromReflect + Typed,
 {
@@ -138,7 +175,7 @@ impl <T, const N: usize> GetTypeDependencies for [T; N] where
 
 macro_rules! register_tuple_dependencies {
     ($($param:ident),*) => {
-        impl <$($param),*> $crate::bindings::GetTypeDependencies for ($($param,)*) where 
+        impl <$($param),*> $crate::bindings::GetTypeDependencies for ($($param,)*) where
             $(
                 $param: GetTypeDependencies,
                 <$param>::Underlying: FromReflect + Typed + GetTypeRegistration,
@@ -155,7 +192,6 @@ macro_rules! register_tuple_dependencies {
     };
 }
 
-
 variadics_please::all_tuples!(register_tuple_dependencies, 1, 14, T);
 
 /// A trait collecting type dependency information for a whole function. Used to register everything used by a function with the type registry
@@ -166,7 +202,7 @@ pub trait GetFunctionTypeDependencies<Marker> {
 
 macro_rules! impl_script_function_type_dependencies{
     ($( $param:ident ),* ) => {
-        impl<F, $( $param,)* O > GetFunctionTypeDependencies<fn($($param),*) -> O> for F where 
+        impl<F, $( $param,)* O > GetFunctionTypeDependencies<fn($($param),*) -> O> for F where
             O: GetTypeDependencies,
             F: Fn( $( $param ),* ) -> O,
             $(
