@@ -2,7 +2,7 @@
 
 use crate::bindings::function::arg_meta::ArgMeta;
 use crate::bindings::function::namespace::Namespace;
-use bevy::reflect::Reflect;
+use bevy_reflect::Reflect;
 use std::{any::TypeId, borrow::Cow};
 
 use super::typed_through::{ThroughTypeInfo, TypedThrough};
@@ -195,14 +195,14 @@ variadics_please::all_tuples!(impl_documentable, 0, 13, T);
 
 #[cfg(test)]
 mod test {
-	use crate::{
-		bindings::function::from::{Mut, Ref, Val},
-		docgen::typed_through::UntypedWrapperKind,
-	};
+    use crate::{
+        bindings::function::from::{Mut, Ref, Val},
+        docgen::typed_through::UntypedWrapperKind,
+    };
 
-	use super::*;
+    use super::*;
 
-	#[test]
+    #[test]
     fn test_get_function_info() {
         fn test_fn(a: i32, b: f32) -> f64 {
             (a as f64) + (b as f64)
