@@ -2,15 +2,17 @@
 
 use std::sync::Arc;
 
+use ::{bevy_asset::Handle, bevy_ecs::entity::Entity, bevy_reflect::Reflect};
+use bevy_ecs::event::Event;
+use parking_lot::Mutex;
+
 use crate::{
+    IntoScriptPluginParams,
     asset::Language,
     bindings::script_value::ScriptValue,
     error::ScriptError,
     script::{ScriptAttachment, ScriptContext, ScriptId},
-    IntoScriptPluginParams,
 };
-use bevy::{asset::Handle, ecs::entity::Entity, prelude::Event, reflect::Reflect};
-use parking_lot::Mutex;
 
 /// A script event
 #[derive(Event, Debug, Clone, PartialEq, Eq)]
@@ -408,20 +410,20 @@ static FORBIDDEN_KEYWORDS: [&str; 82] = [
 mod test {
     use std::sync::Arc;
 
-    use bevy::{
-        asset::{AssetId, AssetIndex, Handle},
-        ecs::entity::Entity,
+    use ::{
+        bevy_app::{App, Plugin},
+        bevy_asset::{AssetId, AssetIndex, Handle},
+        bevy_ecs::entity::Entity,
     };
     use parking_lot::Mutex;
     use test_utils::make_test_plugin;
 
+    use super::FORBIDDEN_KEYWORDS;
     use crate::{
         bindings::ScriptValue,
         event::Recipients,
         script::{ContextPolicy, ScriptAttachment, ScriptContext},
     };
-
-    use super::FORBIDDEN_KEYWORDS;
 
     #[test]
     fn test_invalid_strings() {
