@@ -1,11 +1,9 @@
 //! "Runtime" here refers to the execution evironment of scripts. This might be the VM executing bytecode or the interpreter executing source code.
 //! The important thing is that there is only one runtime which is used to execute all scripts of a particular type or `context`.
 
-use crate::{error::ScriptError, IntoScriptPluginParams};
-use bevy::{
-    ecs::system::ResMut,
-    prelude::{Res, Resource},
-};
+use ::bevy_ecs::{resource::Resource, system::Res, system::ResMut};
+
+use crate::{IntoScriptPluginParams, error::ScriptError};
 
 /// A trait that all script runtimes must implement.
 pub trait Runtime: Default + 'static + Send + Sync {}
