@@ -1,12 +1,13 @@
 //! The library crate for the mdbook LAD preprocessor.
 #![allow(missing_docs)]
 
+use std::sync::OnceLock;
+
 use mdbook::{
     errors::Error,
     preprocess::{Preprocessor, PreprocessorContext},
 };
 use sections::{Section, SectionData};
-use std::sync::OnceLock;
 mod argument_visitor;
 mod markdown;
 mod sections;
@@ -105,19 +106,19 @@ impl Preprocessor for LADPreprocessor {
                     .iter()
                     .enumerate()
                     .filter_map(|(idx, item)| {
-                        if let mdbook::BookItem::Chapter(chapter) = item {
-                            if LADPreprocessor::is_lad_file(chapter) {
-                                match LADPreprocessor::process_lad_chapter(
-                                    context,
-                                    chapter,
-                                    Some(parent),
-                                    idx,
-                                ) {
-                                    Ok(new_chapter) => return Some((idx, new_chapter)),
-                                    Err(e) => {
-                                        errors.push(e);
-                                        return None;
-                                    }
+                        if let mdbook::BookItem::Chapter(chapter) = item
+                            && LADPreprocessor::is_lad_file(chapter)
+                        {
+                            match LADPreprocessor::process_lad_chapter(
+                                context,
+                                chapter,
+                                Some(parent),
+                                idx,
+                            ) {
+                                Ok(new_chapter) => return Some((idx, new_chapter)),
+                                Err(e) => {
+                                    errors.push(e);
+                                    return None;
                                 }
                             }
                         }
@@ -168,7 +169,7 @@ impl Preprocessor for LADPreprocessor {
 
         if !errors.is_empty() {
             // return on first error
-            for error in errors {
+            if let Some(error) = errors.into_iter().next() {
                 log::error!("{error}");
                 Err(error)?;
             }
