@@ -1,14 +1,16 @@
-use crate::{
-    argument_visitor::MarkdownArgumentVisitor,
-    markdown::{markdown_substring, IntoMarkdown, Markdown, MarkdownBuilder, TableBuilder},
-    markdown_vec,
-};
+use std::{borrow::Cow, collections::HashSet, path::PathBuf};
+
 use ladfile::{
     ArgumentVisitor, LadArgument, LadBMSPrimitiveKind, LadFile, LadFunction, LadInstance, LadType,
     LadTypeId, LadTypeKind, LadTypeLayout,
 };
 use mdbook::book::{Chapter, SectionNumber};
-use std::{borrow::Cow, collections::HashSet, path::PathBuf};
+
+use crate::{
+    argument_visitor::MarkdownArgumentVisitor,
+    markdown::{IntoMarkdown, Markdown, MarkdownBuilder, TableBuilder, markdown_substring},
+    markdown_vec,
+};
 
 fn print_type(ladfile: &LadFile, type_: &LadTypeId) -> String {
     let mut visitor = MarkdownArgumentVisitor::new(ladfile);
@@ -267,8 +269,9 @@ impl<'a> Section<'a> {
                             builder.row(markdown_vec![
                                 Markdown::new_paragraph("Global Functions").code(),
                                 Markdown::Link {
-                                    text:
-                                        Box::new("Documents all the global functions present in the bindings"),
+                                    text: Box::new(
+                                        "Documents all the global functions present in the bindings"
+                                    ),
                                     url: format!(
                                         "./{}/functions.md",
                                         linkify_filename(title.clone())
@@ -279,7 +282,9 @@ impl<'a> Section<'a> {
                             builder.row(markdown_vec![
                                 Markdown::new_paragraph("Globals").code(),
                                 Markdown::Link {
-                                    text: Box::new("Documents all global variables present in the bindings"),
+                                    text: Box::new(
+                                        "Documents all global variables present in the bindings"
+                                    ),
                                     url: format!(
                                         "./{}/globals.md",
                                         linkify_filename(title.clone())
