@@ -1,14 +1,17 @@
-use super::reference::LuaReflectReference;
+use std::{
+    collections::VecDeque,
+    ops::{Deref, DerefMut},
+};
+
 use bevy_mod_scripting_core::{
     asset::Language,
     bindings::{function::script_function::FunctionCallContext, script_value::ScriptValue},
     error::InteropError,
 };
+use bevy_platform::collections::HashMap;
 use mlua::{FromLua, IntoLua, Value, Variadic};
-use std::{
-    collections::{HashMap, VecDeque},
-    ops::{Deref, DerefMut},
-};
+
+use super::reference::LuaReflectReference;
 
 #[derive(Debug, Clone)]
 /// A wrapper around a [`ScriptValue`] that implements [`FromLua`] and [`IntoLua`]
@@ -112,7 +115,7 @@ impl FromLua for LuaScriptValue {
                     from: value.type_name(),
                     to: "ScriptValue".to_owned(),
                     message: Some("unsupported value type".to_owned()),
-                })
+                });
             }
         }
         .into())
@@ -160,7 +163,7 @@ impl IntoLua for LuaScriptValue {
                 Value::Table(table)
             }
             ScriptValue::Map(map) => {
-                let hashmap: HashMap<String, Value> = map
+                let hashmap: std::collections::HashMap<String, Value> = map
                     .into_iter()
                     .map(|(k, v)| Ok((k, LuaScriptValue::from(v).into_lua(lua)?)))
                     .collect::<Result<_, mlua::Error>>()?;
