@@ -1,18 +1,21 @@
-use super::script_value::{FromDynamic, FunctionWithReceiver, IntoDynamic, RHAI_CALLER_CONTEXT};
+use std::{
+    any::TypeId,
+    ops::{Deref, DerefMut},
+};
+
 use bevy_mod_scripting_core::{
     bindings::{
+        ReflectReference, ThreadWorldContainer, WorldContainer,
         function::script_function::DynamicScriptFunctionMut, pretty_print::DisplayWithWorld,
-        script_value::ScriptValue, ReflectReference, ThreadWorldContainer, WorldContainer,
+        script_value::ScriptValue,
     },
     error::InteropError,
     reflection_extensions::TypeIdExtensions,
 };
 use rhai::{CustomType, Dynamic, EvalAltResult};
-use std::{
-    any::TypeId,
-    ops::{Deref, DerefMut},
-};
 use strum::VariantNames;
+
+use super::script_value::{FromDynamic, FunctionWithReceiver, IntoDynamic, RHAI_CALLER_CONTEXT};
 
 #[derive(Debug, strum::EnumString, strum::VariantNames, Clone)]
 /// A list of reserved keywords in Rhai
@@ -304,7 +307,7 @@ impl CustomType for RhaiReflectReference {
                         {
                             Ok(func) => {
                                 return FunctionWithReceiver::curry(func, self_.clone().into())
-                                    .into_dynamic()
+                                    .into_dynamic();
                             }
                             Err(string) => ScriptValue::String(string),
                         }
