@@ -1,3 +1,5 @@
+use std::str::FromStr;
+
 use bevy_mod_scripting_core::{
     asset::Language,
     bindings::{
@@ -7,7 +9,6 @@ use bevy_mod_scripting_core::{
     error::InteropError,
 };
 use rhai::{Dynamic, EvalAltResult, FnPtr, Map, NativeCallContext};
-use std::str::FromStr;
 
 use super::reference::RhaiReflectReference;
 
@@ -124,7 +125,7 @@ impl IntoDynamic for ScriptValue {
                     "Interop error in rhai script".to_string(),
                     interop_error.into(),
                 )
-                .into())
+                .into());
             }
         })
     }
