@@ -12,8 +12,8 @@ macro_rules! make_test_plugin {
             }
         }
 
-        impl bevy::app::Plugin for TestPlugin {
-            fn build(&self, app: &mut bevy::app::App) {
+        impl Plugin for TestPlugin {
+            fn build(&self, app: &mut App) {
                 self.0.build(app);
             }
         }
@@ -29,16 +29,42 @@ macro_rules! make_test_plugin {
                     invocations: vec![].into(),
                 }
             }
+
+            fn handler() -> $ident::HandlerFn<Self> {
+                (|args, context_key, callback, script_ctxt, pre_handling_initializers, runtime| {
+                    runtime
+                        .invocations
+                        .lock()
+                        .push((context_key.entity(), Some(context_key.script().id())));
+                    Ok($ident::bindings::script_value::ScriptValue::Unit)
+                }) as $ident::HandlerFn<Self>
+            }
+
+            fn context_loader() -> $ident::ContextLoadFn<Self> {
+                (|attachment, content, context_initializers, pre_handling_initializers, runtime| {
+                    Ok(TestContext {
+                        invocations: vec![],
+                    })
+                })
+            }
+
+            fn context_reloader() -> $ident::ContextReloadFn<Self> {
+                (|attachment,
+                  content,
+                  previous_context,
+                  context_initializers,
+                  pre_handling_initializers,
+                  runtime| {
+                    previous_context.invocations.clear();
+                    Ok(())
+                })
+            }
         }
 
         #[derive(Default, std::fmt::Debug)]
         struct TestRuntime {
-            pub invocations: parking_lot::Mutex<
-                Vec<(
-                    Option<bevy::prelude::Entity>,
-                    Option<$ident::script::ScriptId>,
-                )>,
-            >,
+            pub invocations:
+                parking_lot::Mutex<Vec<(Option<Entity>, Option<$ident::script::ScriptId>)>>,
         }
 
         #[derive(Default, std::fmt::Debug, Clone)]
