@@ -1,11 +1,13 @@
-use bevy::ecs::reflect::AppTypeRegistry;
-use bevy::{app::App, DefaultPlugins};
+use bevy::{DefaultPlugins, app::App, ecs::reflect::AppTypeRegistry};
 use bevy_mod_scripting::ScriptFunctionsPlugin;
-use bevy_mod_scripting_core::bindings::function::script_function::AppScriptFunctionRegistry;
-use bevy_mod_scripting_core::bindings::globals::core::CoreScriptGlobalsPlugin;
-use bevy_mod_scripting_core::bindings::globals::AppScriptGlobalsRegistry;
-use bevy_mod_scripting_core::BMSScriptingInfrastructurePlugin;
-use ladfile_builder::plugin::{generate_lad_file, LadFileSettings, ScriptingDocgenPlugin};
+use bevy_mod_scripting_core::{
+    BMSScriptingInfrastructurePlugin,
+    bindings::{
+        function::script_function::AppScriptFunctionRegistry,
+        globals::{AppScriptGlobalsRegistry, core::CoreScriptGlobalsPlugin},
+    },
+};
+use ladfile_builder::plugin::{LadFileSettings, ScriptingDocgenPlugin, generate_lad_file};
 
 fn main() -> std::io::Result<()> {
     let mut app = App::new();
